split files up a bit more to enhance readability
